requirements.txt
python-vlc
requests
jsons
